refactor: physics 관련 컴포넌트 모듈화
